4.21

+++ controllers/blogs.js
@@ -12,13 +12,26 @@ const User = require('../models/user')
 // }
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({}).populate('user', {username: 1 , name: 1 })
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify((request.token), process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+  
+  if (user._id.toString() === blog.user.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).json("Blog deleted succesfully").end()
+  } else {
+    response.status(400).json("Unauthorised to delete").end()
+  }
+
 })
 
 
@@ -61,7 +74,7 @@ blogsRouter.post('/', async (request, response, next) => {
     const saveBlog = await blog.save()
     user.blogs = user.blogs.concat(saveBlog._id)
     await user.save()
-    
+
     response.status(201).json(saveBlog)
 
   } catch (exception) {
